move some files in common
