+++ lib/appium_lib/common/patch.rb
