+++ lib/appium_lib/common/search_context.rb
@@ -1,6 +1,7 @@
-require_relative 'core/search_context'
-require_relative 'android/search_context'
-require_relative 'ios/search_context'
+require_relative '../core/search_context'
+require_relative '../android/search_context'
+require_relative '../ios/search_context'
+
 module Appium
   class Driver
     module SearchContext
